build errors

+++ impl/src/tooling/interpreter/common.h
@@ -104,7 +104,7 @@ typedef uint64_t RCMeta;
 
 #define GC_RC_CLEAR ((uint64_t)0)
 #define GC_RC_ETERNAL_INIT ((uint64_t)1)
-#define GC_RC_MARK_FROM_ROOT ((uint64_t)(1 << 60))
+#define GC_RC_MARK_FROM_ROOT ((uint64_t)(((uint64_t)1) << 60))
 
 #define GC_COUNT_GET_MASK 0xFFFFFFFFFFFF
 #define GC_MARK_GET_MASK 0xFFFF000000000000
